Use MUnit in SwaggerFormatsSuite
