+++ project/Dependencies.scala
@@ -26,6 +26,8 @@ object Dependencies {
   lazy val shapeless           = "com.chuusai"                %% "shapeless"             % "2.3.3"
   lazy val scalaXml            = "org.scala-lang.modules"     %% "scala-xml"             % "2.0.0"
   lazy val swaggerUi           = "org.webjars"                 % "swagger-ui"            % "3.46.0"
+  lazy val munit               = "org.scalameta"              %% "munit"                 % "0.7.26"
+  lazy val scalacheckMunit     = "org.scalameta"              %% "munit-scalacheck"      % munit.revision
 
   lazy val specs2              = Seq("org.specs2"              %% "specs2-core"          % specs2Version % "test",
                                      "org.specs2"              %% "specs2-scalacheck"    % specs2Version % "test" )
@@ -46,7 +48,7 @@ object Dependencies {
 
     http4sCirce % "test",
     circeParser % "test",
-    circeGeneric % "test",
+    circeGeneric % "test"
   )
 
   lazy val swaggerUiDeps = libraryDependencies ++= Seq(swaggerUi)
@@ -60,6 +62,6 @@ object Dependencies {
     http4sCirce,
     http4sXmlInstances,
     logbackClassic, 
-    uadetector,
+    uadetector
   )
 }
